+++ src/components/TaskSection.tsx
@@ -6,13 +6,14 @@ import {
 } from 'recoil';
 import { groupBy } from 'lodash-es';
 import { TaskEntityObject } from '../models/TaskEntity';
-import { tasksState } from '../state/tasks';
+import { filterdTasksState } from '../state/tasks';
 import { themeStyleState } from '../state/theme';
 import { visibleState } from '../state/visible';
 import TaskItem from './TaskItem';
 import { settingsState } from '../state/settings';
-import { openTaskPage } from '../api';
+import { openTask, openTaskPage } from '../api';
 import { inputState } from '../state/input';
+import { ChevronsRight } from 'tabler-icons-react';
 
 export enum GroupBy {
   Page,
@@ -30,7 +31,7 @@ const TaskSection: React.FC<ITaskSectionProps> = (props) => {
   const { title, query } = props;
   const [tasks, setTasks] = useState<TaskEntityObject[]>([]);
   const visible = useRecoilValue(visibleState);
-  const tasksLoadable = useRecoilValueLoadable(tasksState(query));
+  const tasksLoadable = useRecoilValueLoadable(filterdTasksState(query));
   const themeStyle = useRecoilValue(themeStyleState);
   const { openInRightSidebar } = useRecoilValue(settingsState);
   const input = useRecoilValue(inputState);
@@ -76,18 +77,35 @@ const TaskSection: React.FC<ITaskSectionProps> = (props) => {
     }
   }, [props.groupBy, tasks]);
 
+  const openTaskGroups = React.useCallback(() => {
+    tasks.forEach((task) => {
+      openTask(task, {
+        openInRightSidebar: true,
+      });
+    });
+    window.logseq.hideMainUI();
+  }, [tasks]);
+
   if (tasks.length === 0) {
     return null;
   }
 
   return (
     <div className="py-1">
-      <h2
-        className="py-1 text-blue-400"
-        style={{ color: themeStyle.sectionTitleColor }}
-      >
-        {title}
-      </h2>
+      <div className="flex flex-row justify-between items-center">
+        <h2
+          className="py-1 mr-2 text-blue-400"
+          style={{ color: themeStyle.sectionTitleColor }}
+        >
+          {title}
+        </h2>
+        <div className="pl-2 pr-1" onClick={openTaskGroups}>
+          <ChevronsRight
+            size={20}
+            className="stroke-gray-300 cursor-pointer"
+          />
+        </div>
+      </div>
       <div>
         {(Object.entries(taskGroups) ?? []).map(([name, tasks]) => {
           const [{ page }] = tasks;