+++ src/querys/scheduled.ts
@@ -1,17 +1,24 @@
-import dayjs from 'dayjs';
+import dayjs, { Dayjs } from 'dayjs';
 
-export default function getScheduledTaskQuery() {
-  const today = dayjs().format('YYYYMMDD');
+export default function getScheduledTaskQuery(
+  startDate: Dayjs | Date = new Date(),
+) {
+  const start = dayjs(startDate).format('YYYYMMDD');
   const query = `
     [:find (pull ?b [*])
      :where
      [?b :block/marker ?marker]
-     [(contains? #{"NOW" "LATER" "TODO" "DOING" "WAITING"} ?marker)]
+     [(contains? #{"NOW" "LATER" "TODO" "DOING"} ?marker)]
      [?b :block/page ?p]
      (or
-       [?b :block/scheduled ?d]
-       [?b :block/deadline ?d])
-     [(> ?d ${today})]]
+       (or
+         [?b :block/scheduled ?d]
+         [?b :block/deadline ?d])
+       (and
+         [?p :block/journal? true]
+         [?p :block/journal-day ?d]))
+     [(> ?d ${start})]]
   `;
+
   return query;
 }