+++ src/state/settings.ts
@@ -1,7 +1,14 @@
 import { atom, AtomEffect } from 'recoil';
+import { TaskMarker, TaskPriority } from '../models/TaskEntity';
+import settings from '../settings';
 
 interface IPluginSettings {
   hotkey: string;
+  defaultMarker: TaskMarker;
+  customMarkers: string;
+  defaultPriority: TaskPriority;
+  showNextNDaysTask: boolean;
+  numberOfNextNDays: number;
   lightPrimaryBackgroundColor: string;
   lightSecondaryBackgroundColor: string;
   darkPrimaryBackgroundColor: string;
@@ -10,17 +17,6 @@ interface IPluginSettings {
   openInRightSidebar: boolean;
   whereToPlaceNewTask: string;
 }
-
-const DEFAULT_SETTINGS = {
-  hotkey: 'mod+shift+t',
-  sectionTitleColor: '#0a0a0a',
-  lightPrimaryBackgroundColor: '#ffffff',
-  lightSecondaryBackgroundColor: '#f7f7f7',
-  darkPrimaryBackgroundColor: '#002B37',
-  darkSecondaryBackgroundColor: '#106ba3',
-  openInRightSidebar: false,
-  whereToPlaceNewTask: '',
-};
 
 const settingsChangedEffect: AtomEffect<IPluginSettings> = ({ setSelf }) => {
   setSelf({ ...logseq.settings } as unknown as IPluginSettings);
@@ -32,6 +28,6 @@ const settingsChangedEffect: AtomEffect<IPluginSettings> = ({ setSelf }) => {
 
 export const settingsState = atom<IPluginSettings>({
   key: 'settings',
-  default: DEFAULT_SETTINGS,
+  default: settings.reduce((result, item) => ({ ...result, [item.key]: item.default }), {}) as IPluginSettings,
   effects: [settingsChangedEffect],
 });