+++ src/state/user-configs.ts
@@ -1,5 +1,10 @@
 import { AppUserConfigs } from '@logseq/libs/dist/LSPlugin';
-import { atom, AtomEffect } from 'recoil';
+import { atom, AtomEffect, selector } from 'recoil';
+import { logseq as plugin } from '../../package.json';
+import { TaskMarker } from '../models/TaskEntity';
+import { settingsState } from './settings';
+
+export const USER_CONFIGS_KEY = `${plugin.id}#userConfigs`;
 
 export const DEFAULT_USER_CONFIGS: Partial<AppUserConfigs> = {
   preferredLanguage: 'en',
@@ -10,36 +15,47 @@ export const DEFAULT_USER_CONFIGS: Partial<AppUserConfigs> = {
   preferredDateFormat: 'MMM do, yyyy',
 };
 
-function fixPreferredDateFormat(preferredDateFormat: string) {
-  const format = preferredDateFormat
-    .replace('yyyy', 'YYYY')
-    .replace('dd', 'DD')
-    .replace('do', 'Do')
-    .replace('EEEE', 'dddd')
-    .replace('EEE', 'ddd')
-    .replace('EE', 'dd')
-    .replace('E', 'dd');
-  return format;
-}
-
-const updateUserConfigsEffect: AtomEffect<Partial<AppUserConfigs>> = ({
-  setSelf,
-  trigger,
-}) => {
-  if (trigger === 'get') {
-    window.logseq.App.getUserConfigs()
-      .then((configs) => ({
-        ...configs,
-        preferredDateFormat: fixPreferredDateFormat(
-          configs.preferredDateFormat,
-        ),
-      }))
-      .then(setSelf);
-  }
+const themeModeChangeEffect: AtomEffect<AppUserConfigs> = ({ onSet }) => {
+  onSet(({ preferredThemeMode }) => {
+    if (preferredThemeMode === 'dark') {
+      document.documentElement.classList.add('dark');
+      document.documentElement.classList.remove('light');
+    } else {
+      document.documentElement.classList.add('light');
+      document.documentElement.classList.remove('dark');
+    }
+  });
 };
 
-export const userConfigsState = atom({
+const localStorageEffect: AtomEffect<AppUserConfigs> = ({ setSelf, onSet }) => {
+  const savedValue = localStorage.getItem(USER_CONFIGS_KEY);
+  if (savedValue != null) {
+    setSelf(JSON.parse(savedValue));
+  }
+
+  onSet((newValue, _, isReset) => {
+    isReset
+      ? localStorage.removeItem(USER_CONFIGS_KEY)
+      : localStorage.setItem(USER_CONFIGS_KEY, JSON.stringify(newValue));
+  });
+};
+
+export const userConfigsState = atom<AppUserConfigs>({
   key: 'userConfigs',
-  default: DEFAULT_USER_CONFIGS,
-  effects: [updateUserConfigsEffect],
+  default: DEFAULT_USER_CONFIGS as AppUserConfigs,
+  effects: [localStorageEffect, themeModeChangeEffect],
+});
+
+export const taskMarkersState = selector<(TaskMarker | string)[]>({
+  key: 'taskMarkers',
+  get: ({ get }) => {
+    const { preferredWorkflow } = get(userConfigsState);
+    const settings = get(settingsState);
+    const customMarkers =
+      settings.customMarkers === '' ? [] : settings.customMarkers.split(',');
+    if (preferredWorkflow === 'now') {
+      return [TaskMarker.LATER, TaskMarker.NOW, ...customMarkers];
+    }
+    return [TaskMarker.TODO, TaskMarker.DOING, ...customMarkers];
+  },
 });